Sensorslib: Refactor `SensorsConsumerRn`
This commit will rename `SensorsConsumerRn` to `SensorEventConsumerRn` and replace all
of it's usages throuhout the code.

This commit will also move the `SensorsConsumer` interface in it's own class and rename
it `SensorDataHandler`
