+++ sensorslib/src/test/java/com/handmonitor/sensorslib/SensorEventConsumerTest.kt
@@ -1,25 +1,22 @@
 package com.handmonitor.sensorslib
 
-import android.util.Log
 import com.google.common.truth.Truth.assertThat
 import io.mockk.every
 import io.mockk.impl.annotations.MockK
 import io.mockk.junit5.MockKExtension
-import io.mockk.mockkStatic
 import io.mockk.verify
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 
 @ExtendWith(MockKExtension::class)
-class SensorsConsumerTest {
+class SensorEventConsumerTest {
     @MockK
     private lateinit var mData: SensorSharedData
 
     @BeforeEach
     fun setup() {
-        mockkStatic(Log::class)
-        every { Log.d(any(), any()) } returns 0
+        mockLog()
     }
 
     @Test
@@ -27,13 +24,13 @@ class SensorsConsumerTest {
         val mockData = floatArrayOf(0.0f, 1.1f, 2.2f, 3.3f)
         every { mData.getData() } returns mockData
 
-        val impl = object : SensorsConsumer {
+        val impl = object : SensorDataHandler {
             override fun onNewData(data: FloatArray) {
                 assertThat(data).isEqualTo(mockData)
                 throw InterruptedException()
             }
         }
-        val runnable = SensorsConsumerRn(mData, impl)
+        val runnable = SensorEventConsumerRn(mData, impl)
         runnable.run()
 
         verify {
